+++ src/lib.rs
@@ -1,17 +1,14 @@
 pub extern crate gazetteer_entity_parser;
 
-pub use snips_nlu_ontology::*;
-
 pub use builtin_entity_parser::*;
 pub use conversion::*;
 pub use gazetteer_parser::*;
-
-pub mod errors;
+pub use snips_nlu_ontology::*;
 
 mod builtin_entity_parser;
 mod conversion;
+pub mod errors;
 mod gazetteer_parser;
 #[cfg(test)]
 mod test_utils;
 mod utils;
-