+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [Unreleased]
+### Added
+- Bump `gazetteer-entity-parser` to `0.7.1` in order to handle license files [#28](https://github.com/snipsco/snips-nlu-parsers/pull/28)
+
 ## [0.3.0]
 ### Changed
 - Bump `snips-nlu-ontology` to `0.65.0` [#26](https://github.com/snipsco/snips-nlu-parsers/pull/26)
@@ -40,6 +44,7 @@ All notable changes to this project will be documented in this file.
 - bump `snips-nlu-ontology` to `0.63.0`
 - re-export `gazetteer-entity-parser` crate
 
+[Unreleased]: https://github.com/snipsco/snips-nlu-parsers/compare/0.3.0...HEAD
 [0.3.0]: https://github.com/snipsco/snips-nlu-parsers/compare/0.2.3...0.3.0
 [0.2.3]: https://github.com/snipsco/snips-nlu-parsers/compare/0.2.2...0.2.3
 [0.2.2]: https://github.com/snipsco/snips-nlu-parsers/compare/0.2.1...0.2.2