+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
-## [Unreleased]
+## [0.3.1]
 ### Added
+- Add Builtin Entity support per Language [#29](https://github.com/snipsco/snips-nlu-parsers/pull/29)
 - Bump `gazetteer-entity-parser` to `0.7.1` in order to handle license files [#28](https://github.com/snipsco/snips-nlu-parsers/pull/28)
 
 ## [0.3.0]
@@ -44,7 +45,7 @@ All notable changes to this project will be documented in this file.
 - bump `snips-nlu-ontology` to `0.63.0`
 - re-export `gazetteer-entity-parser` crate
 
-[Unreleased]: https://github.com/snipsco/snips-nlu-parsers/compare/0.3.0...HEAD
+[0.3.1]: https://github.com/snipsco/snips-nlu-parsers/compare/0.3.0...0.3.1
 [0.3.0]: https://github.com/snipsco/snips-nlu-parsers/compare/0.2.3...0.3.0
 [0.2.3]: https://github.com/snipsco/snips-nlu-parsers/compare/0.2.2...0.2.3
 [0.2.2]: https://github.com/snipsco/snips-nlu-parsers/compare/0.2.1...0.2.2